Code is changed for CC as a source of fund

+++ src/main/java/com/mashreq/transfercoreservice/fundtransfer/dto/QRDealDetails.java
@@ -19,7 +19,7 @@ public class QRDealDetails implements Serializable {
 
     private Integer totalLimitAmount;
 
-    private Integer utilizedLimitAmount;
+    private Float utilizedLimitAmount;
 
     private Date eligibilityStartDate;
